Creater resources.ts and moved some functions there
