+++ backend/src/routes/course.ts
@@ -1,7 +1,6 @@
 import express from "express";
 import {
   createCourse,
-  createResource,
   generateResources,
   getUserCourses,
   saveResources,
